+++ src/azure-cli/azure/cli/command_modules/sql/custom.py
@@ -46,7 +46,6 @@ from ._util import (
     get_sql_managed_instances_operations
 )
 
-
 logger = get_logger(__name__)
 
 ###############################################
@@ -96,6 +95,27 @@ def _any_sku_values_specified(sku):
     '''
 
     return any(val for key, val in sku.__dict__.items())
+
+
+def _compute_model_matches(sku_name, compute_model):
+    '''
+    Returns True if sku name matches the compute model.
+    Please update is function if compute_model has more than 2 enums.
+    '''
+
+    if (_is_serverless_slo(sku_name) and compute_model == ComputeModelType.serverless):
+        return True
+    if (not _is_serverless_slo(sku_name) and compute_model != ComputeModelType.serverless):
+        return True
+    return False
+
+
+def _is_serverless_slo(sku_name):
+    '''
+    Returns True if the sku name is a serverless sku.
+    '''
+
+    return "_S_" in sku_name
 
 
 def _get_default_server_version(location_capabilities):
@@ -225,7 +245,7 @@ def _find_family_capability(sku, supported_families):
         return _get_default_capability(supported_families)
 
 
-def _find_performance_level_capability(sku, supported_service_level_objectives, allow_reset_family):
+def _find_performance_level_capability(sku, supported_service_level_objectives, allow_reset_family, compute_model=None):
     '''
     Finds the DB or elastic pool performance level (i.e. service objective) in the
     collection of supported service objectives that matches the requested sku's
@@ -235,8 +255,8 @@ def _find_performance_level_capability(sku, supported_service_level_objectives, 
     objective.
     '''
 
-    logger.debug('_find_performance_level_capability: %s, %s, allow_reset_family: %s',
-                 sku, supported_service_level_objectives, allow_reset_family)
+    logger.debug('_find_performance_level_capability: %s, %s, allow_reset_family: %s, compute_model: %s',
+                 sku, supported_service_level_objectives, allow_reset_family, compute_model)
 
     if sku.capacity:
         try:
@@ -245,7 +265,8 @@ def _find_performance_level_capability(sku, supported_service_level_objectives, 
             return next(slo for slo in supported_service_level_objectives
                         if ((slo.sku.family == sku.family) or
                             (slo.sku.family is None and allow_reset_family)) and
-                        int(slo.sku.capacity) == int(sku.capacity))
+                        int(slo.sku.capacity) == int(sku.capacity) and
+                        _compute_model_matches(slo.sku.name, compute_model))
         except StopIteration:
             if allow_reset_family:
                 raise CLIError(
@@ -281,7 +302,8 @@ def _db_elastic_pool_update_sku(
         tier,
         family,
         capacity,
-        find_sku_from_capabilities_func):
+        find_sku_from_capabilities_func,
+        compute_model=None):
     '''
     Updates the sku of a DB or elastic pool.
     '''
@@ -319,12 +341,18 @@ def _db_elastic_pool_update_sku(
     if capacity:
         instance.sku.capacity = capacity
 
+    # Wipe out sku name if serverless vs provisioned db offerings changed,
+    # only if sku name has not be wiped by earlier logic, and new compute model has been requested.
+    if instance.sku.name and compute_model:
+        if not _compute_model_matches(instance.sku.name, compute_model):
+            instance.sku.name = None
+
     # If sku name was wiped out by any of the above, resolve the requested sku name
     # using capabilities.
     if not instance.sku.name:
         instance.sku = find_sku_from_capabilities_func(
             cmd.cli_ctx, instance.location, instance.sku,
-            allow_reset_family=allow_reset_family)
+            allow_reset_family=allow_reset_family, compute_model=compute_model)
 
 
 def _get_tenant_id():
@@ -393,6 +421,12 @@ class ClientAuthenticationType(Enum):
 class FailoverPolicyType(Enum):
     automatic = 'Automatic'
     manual = 'Manual'
+
+
+class ComputeModelType(str, Enum):
+
+    provisioned = "Provisioned"
+    serverless = "Serverless"
 
 
 def _get_server_dns_suffx(cli_ctx):
@@ -546,7 +580,7 @@ class DatabaseIdentity():  # pylint: disable=too-few-public-methods
             quote(self.database_name))
 
 
-def _find_db_sku_from_capabilities(cli_ctx, location, sku, allow_reset_family=False):
+def _find_db_sku_from_capabilities(cli_ctx, location, sku, allow_reset_family=False, compute_model=None):
     '''
     Given a requested sku which may have some properties filled in
     (e.g. tier and capacity), finds the canonical matching sku
@@ -582,7 +616,8 @@ def _find_db_sku_from_capabilities(cli_ctx, location, sku, allow_reset_family=Fa
     # Find performance level capability, based on requested sku properties
     performance_level_capability = _find_performance_level_capability(
         sku, edition_capability.supported_service_level_objectives,
-        allow_reset_family=allow_reset_family)
+        allow_reset_family=allow_reset_family,
+        compute_model=compute_model)
 
     # Ideally, we would return the sku object from capability (`return performance_level_capability.sku`).
     # However not all db create modes support using `capacity` to find slo, so instead we put
@@ -637,6 +672,13 @@ def _db_dw_create(
     Handles common concerns such as setting location and sku properties.
     '''
 
+    # This check needs to be here, because server side logic of
+    # finding a default sku for Serverless is not yet implemented.
+    if kwargs['compute_model'] == ComputeModelType.serverless:
+        if not sku or not sku.tier or not sku.family or not sku.capacity:
+            raise CLIError('When creating a severless database, please pass in edition, '
+                           'family, and capacity parameters through -e -f -c')
+
     # Determine server location
     kwargs['location'] = _get_server_location(
         cli_ctx,
@@ -649,7 +691,11 @@ def _db_dw_create(
 
     # If sku.name is not specified, resolve the requested sku name
     # using capabilities.
-    kwargs['sku'] = _find_db_sku_from_capabilities(cli_ctx, kwargs['location'], sku)
+    kwargs['sku'] = _find_db_sku_from_capabilities(
+        cli_ctx,
+        kwargs['location'],
+        sku,
+        compute_model=kwargs['compute_model'])
 
     # Validate elastic pool id
     kwargs['elastic_pool_id'] = _validate_elastic_pool_id(
@@ -1115,7 +1161,10 @@ def db_update(
         zone_redundant=None,
         tier=None,
         family=None,
-        capacity=None):
+        capacity=None,
+        min_capacity=None,
+        auto_pause_delay=None,
+        compute_model=None):
     '''
     Applies requested parameters to a db resource instance for a DB update.
     '''
@@ -1152,6 +1201,12 @@ def db_update(
         server_name,
         resource_group_name)
 
+    # Finding out requesting compute_model
+    if not compute_model:
+        compute_model = (
+            ComputeModelType.serverless if _is_serverless_slo(instance.sku.name)
+            else ComputeModelType.provisioned)
+
     # Update sku
     _db_elastic_pool_update_sku(
         cmd,
@@ -1160,7 +1215,8 @@ def db_update(
         tier,
         family,
         capacity,
-        find_sku_from_capabilities_func=_find_db_sku_from_capabilities)
+        find_sku_from_capabilities_func=_find_db_sku_from_capabilities,
+        compute_model=compute_model)
 
     # TODO Temporary workaround for elastic pool sku name issue
     if instance.elastic_pool_id:
@@ -1175,6 +1231,15 @@ def db_update(
 
     if zone_redundant is not None:
         instance.zone_redundant = zone_redundant
+
+    #####
+    # Set other (serverless related) properties
+    #####
+    if min_capacity:
+        instance.min_capacity = min_capacity
+
+    if auto_pause_delay:
+        instance.auto_pause_delay = auto_pause_delay
 
     return instance
 
@@ -1514,7 +1579,7 @@ def dw_resume(
 ###############################################
 
 
-def _find_elastic_pool_sku_from_capabilities(cli_ctx, location, sku, allow_reset_family=False):
+def _find_elastic_pool_sku_from_capabilities(cli_ctx, location, sku, allow_reset_family=False, compute_model=None):
     '''
     Given a requested sku which may have some properties filled in
     (e.g. tier and capacity), finds the canonical matching sku
@@ -1549,7 +1614,8 @@ def _find_elastic_pool_sku_from_capabilities(cli_ctx, location, sku, allow_reset
     # Find performance level capability, based on requested sku properties
     performance_level_capability = _find_performance_level_capability(
         sku, edition_capability.supported_elastic_pool_performance_levels,
-        allow_reset_family=allow_reset_family)
+        allow_reset_family=allow_reset_family,
+        compute_model=compute_model)
 
     # Copy sku object from capability
     result = performance_level_capability.sku